+++ src/gulp/build.js
@@ -106,7 +106,7 @@ export default function (opts) {
    */
   var babelTransform = babel.configure({
     // use experimental es7 class properties
-    optional: [ "es7.classProperties" ]
+    optional: [ "es7.classProperties", "es7.decorators" ]
   });
 
   /**